Delete bsdf.rs
